refactor(cli/network-broker): move `struct`s to `lib.rs`
